# Optimize RmtCommand serialization