--- a/nanoFramework.Hardware.Esp32.Rmt.Benchmarks/SerializeCommandsBenchmark.cs
+++ b/nanoFramework.Hardware.Esp32.Rmt.Benchmarks/SerializeCommandsBenchmark.cs
@@ -94,28 +94,5 @@ namespace nanoFramework.Hardware.Esp32.Rmt.Benchmarks
                 i += 4;
             }
         }
-
-        [Benchmark]
-        public void SerializeCommands_Improved_1540()
-        {
-            var index = 0;
-            var serializedCommands = new byte[RmtCommandsArrayList.Count * 4];
-            for (var i = 0; i < RmtCommandsArrayList.Count; i++)
-            {
-                var command = (RmtCommand)RmtCommandsArrayList[i];
-
-                var highByte1 = (byte)(command.Duration0 >> 8);
-                var lowByte1 = (byte)(command.Duration0 & 0xFF);
-                var highByte2 = (byte)(command.Duration1 >> 8);
-                var lowByte2 = (byte)(command.Duration1 & 0xFF);
-                var level1 = (byte)(command.Level0 ? 0x80 : 0);
-                var level2 = (byte)(command.Level1 ? 0x80 : 0);
-
-                serializedCommands[index++] = lowByte1;
-                serializedCommands[index++] = (byte)(highByte1 | level1);
-                serializedCommands[index++] = lowByte2;
-                serializedCommands[index++] = (byte)(highByte2 | level2);
-            }
-        }
     }
 }