--- a/nanoFramework.Hardware.Esp32.Rmt.Benchmarks/packages.config
+++ b/nanoFramework.Hardware.Esp32.Rmt.Benchmarks/packages.config
@@ -1,0 +1,10 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="nanoFramework.Benchmark" version="1.0.102" targetFramework="netnano1.0" />
+  <package id="nanoFramework.CoreLibrary" version="1.17.1" targetFramework="netnano1.0" />
+  <package id="nanoFramework.Logging" version="1.1.142" targetFramework="netnano1.0" />
+  <package id="nanoFramework.Runtime.Native" version="1.7.9" targetFramework="netnano1.0" />
+  <package id="nanoFramework.System.Collections" version="1.5.62" targetFramework="netnano1.0" />
+  <package id="nanoFramework.System.Diagnostics.Stopwatch" version="1.2.815" targetFramework="netnano1.0" />
+  <package id="nanoFramework.System.Text" version="1.3.29" targetFramework="netnano1.0" />
+</packages>