--- a/nanoFramework.Hardware.Esp32.Rmt.UnitTests/packages.config
+++ b/nanoFramework.Hardware.Esp32.Rmt.UnitTests/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="nanoFramework.CoreLibrary" version="1.17.1" targetFramework="netnano1.0" />
+  <package id="nanoFramework.TestFramework" version="3.0.68" targetFramework="netnano1.0" developmentDependency="true" />
+</packages>