--- a/nanoFramework.Hardware.Esp32.Rmt.sln
+++ b/nanoFramework.Hardware.Esp32.Rmt.sln
@@ -11,6 +11,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "nanoFramework.Hardware.Esp32.Rmt", "nanoFramework.Hardware.Esp32.Rmt\nanoFramework.Hardware.Esp32.Rmt.nfproj", "{D8B1966C-881F-48AB-A821-7B1F83303259}"
 EndProject
+Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "nanoFramework.Hardware.Esp32.Rmt.UnitTests", "nanoFramework.Hardware.Esp32.Rmt.UnitTests\nanoFramework.Hardware.Esp32.Rmt.UnitTests.nfproj", "{C91D222A-AF5B-4BA2-B4D6-F4999031AAF9}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -23,6 +25,12 @@ Global
 		{D8B1966C-881F-48AB-A821-7B1F83303259}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D8B1966C-881F-48AB-A821-7B1F83303259}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D8B1966C-881F-48AB-A821-7B1F83303259}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{C91D222A-AF5B-4BA2-B4D6-F4999031AAF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C91D222A-AF5B-4BA2-B4D6-F4999031AAF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C91D222A-AF5B-4BA2-B4D6-F4999031AAF9}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
+		{C91D222A-AF5B-4BA2-B4D6-F4999031AAF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C91D222A-AF5B-4BA2-B4D6-F4999031AAF9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C91D222A-AF5B-4BA2-B4D6-F4999031AAF9}.Release|Any CPU.Deploy.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE