--- a/nanoFramework.Hardware.Esp32.Rmt.sln
+++ b/nanoFramework.Hardware.Esp32.Rmt.sln
@@ -13,6 +13,8 @@ Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "nanoFramework.Hardware.Esp3
 EndProject
 Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "nanoFramework.Hardware.Esp32.Rmt.UnitTests", "nanoFramework.Hardware.Esp32.Rmt.UnitTests\nanoFramework.Hardware.Esp32.Rmt.UnitTests.nfproj", "{C91D222A-AF5B-4BA2-B4D6-F4999031AAF9}"
 EndProject
+Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "nanoFramework.Hardware.Esp32.Rmt.Benchmarks", "nanoFramework.Hardware.Esp32.Rmt.Benchmarks\nanoFramework.Hardware.Esp32.Rmt.Benchmarks.nfproj", "{7C2BC423-F2D8-41DC-A5AF-90CFD014CA07}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -31,6 +33,12 @@ Global
 		{C91D222A-AF5B-4BA2-B4D6-F4999031AAF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C91D222A-AF5B-4BA2-B4D6-F4999031AAF9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C91D222A-AF5B-4BA2-B4D6-F4999031AAF9}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{7C2BC423-F2D8-41DC-A5AF-90CFD014CA07}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7C2BC423-F2D8-41DC-A5AF-90CFD014CA07}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7C2BC423-F2D8-41DC-A5AF-90CFD014CA07}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
+		{7C2BC423-F2D8-41DC-A5AF-90CFD014CA07}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7C2BC423-F2D8-41DC-A5AF-90CFD014CA07}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7C2BC423-F2D8-41DC-A5AF-90CFD014CA07}.Release|Any CPU.Deploy.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE