<!--- In the TITLE (↑↑↑↑ above ↑↑↑↑ **NOT HERE**) provide a general, short summary of your changes -->
<!--- Please DO NOT use references to other PR's or issues -->

## Description
- Optimize RmtCommand serialization as suggested by @eeegs

## Motivation and Context
Provides a 50% improvement in RmtCommand serialization as suggested by @eeegs here: https://github.com/nanoframework/Home/issues/1540#issuecomment-2649443850

```text
| --------------------------------------------------------------------------------------- |
| MethodName                      | IterationCount | Mean                  | Min  | Max   |
| --------------------------------------------------------------------------------------- |
| SerializeCommands_Original      | 2000           | 9.8049999999999997 ms | 0 ms | 20 ms |
| SerializeCommands_Improved_1540 | 2000           | 4.825 ms              | 0 ms | 10 ms |
| --------------------------------------------------------------------------------------- |
```

## How Has This Been Tested?<!-- (IF APPLICABLE) -->
- Locally
- Unit tests
- Benchmarks

## Screenshots<!-- (IF APPROPRIATE): -->

## Types of changes
<!--- What types of changes does this PR introduce? Put an `x` in all the boxes that apply: -->
- [x] Improvement (non-breaking change that improves a feature, code or algorithm)
- [ ] Bug fix (non-breaking change which fixes an issue with code or algorithm)
- [ ] New feature (non-breaking change which adds functionality to code)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Config and build (change in the configuration and build system, has no impact on code or features)
- [ ] Dependencies (update dependencies and changes associated, has no impact on code or features)
- [x] Unit Tests (add new Unit Test(s) or improved existing one(s), has no impact on code or features)
- [ ] Documentation (changes or updates in the documentation, has no impact on code or features)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
<!--- PLEASE PLEASE PLEASE don't tick all of them just because -->
- [x] My code follows the code style of this project (only if there are changes in source code).
- [ ] My changes require an update to the documentation (there are changes that require the docs website to be updated).
- [ ] I have updated the documentation accordingly (the changes require an update on the docs in this repo).
- [x] I have read the [CONTRIBUTING](https://github.com/nanoframework/.github/blob/main/CONTRIBUTING.md) document.
- [x] I have tested everything locally and all new and existing tests passed (only if there are changes in source code).
- [x] I have added new tests to cover my changes.


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **Chores**
  - Enhanced dependency management across key project modules, ensuring more consistent and reliable performance.
  - Implemented tighter version control and integrity checks to reduce the potential for conflicts and improve overall stability.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->