@josesimoes 2025-02-26
@CoryCharlton any chance you can look into what sonarcloud analysis is pointing? If its not doable, let me know and I'll override the warning.