@CoryCharlton 2025-02-26
> @CoryCharlton any chance you can look into what sonarcloud analysis is pointing? If its not doable, let me know and I'll override the warning.

It was complaining about duplicated code in the benchmark project. I removed the code since it was a duplicate benchmark that is already covered in the latest commit.